requirements.txt
pytest
pytest-cov
opencensus-ext-azure
